# apps/api/openapi.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "Visa Help Center API",
                 "version":  "1.0.0",
                 "description":  "Comprehensive visa application management system API"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Development server"
                    },
                    {
                        "url":  "https://api.boltvisa.com",
                        "description":  "Production server"
                    }
                ],
    "components":  {
                       "responses":  {
                                         "BadRequest":  {
                                                            "description":  "Bad request",
                                                            "content":  {
                                                                            "application/json":  {
                                                                                                     "schema":  {
                                                                                                                    "$ref":  "#/components/schemas/Error"
                                                                                                                }
                                                                                                 }
                                                                        }
                                                        },
                                         "Unauthorized":  {
                                                              "description":  "Unauthorized",
                                                              "content":  {
                                                                              "application/json":  {
                                                                                                       "schema":  {
                                                                                                                      "$ref":  "#/components/schemas/Error"
                                                                                                                  }
                                                                                                   }
                                                                          }
                                                          },
                                         "Conflict":  {
                                                          "description":  "Conflict",
                                                          "content":  {
                                                                          "application/json":  {
                                                                                                   "schema":  {
                                                                                                                  "$ref":  "#/components/schemas/Error"
                                                                                                              }
                                                                                               }
                                                                      }
                                                      }
                                     }
                   },
    "paths":  {
                  "/api/v1/auth/register":  {
                                                "post":  {
                                                             "summary":  "Register a new user",
                                                             "requestBody":  {
                                                                                 "required":  true,
                                                                                 "content":  {
                                                                                                 "application/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "type":  "object",
                                                                                                                                         "required":  [
                                                                                                                                                          "email",
                                                                                                                                                          "password",
                                                                                                                                                          "first_name",
                                                                                                                                                          "last_name"
                                                                                                                                                      ],
                                                                                                                                         "properties":  {
                                                                                                                                                            "email":  {
                                                                                                                                                                          "type":  "string",
                                                                                                                                                                          "format":  "email"
                                                                                                                                                                      },
                                                                                                                                                            "password":  {
                                                                                                                                                                             "type":  "string",
                                                                                                                                                                             "minLength":  8
                                                                                                                                                                         },
                                                                                                                                                            "first_name":  {
                                                                                                                                                                               "type":  "string"
                                                                                                                                                                           },
                                                                                                                                                            "last_name":  {
                                                                                                                                                                              "type":  "string"
                                                                                                                                                                          }
                                                                                                                                                        }
                                                                                                                                     }
                                                                                                                      }
                                                                                             }
                                                                             },
                                                             "responses":  {
                                                                               "201":  {
                                                                                           "description":  "User created successfully",
                                                                                           "content":  {
                                                                                                           "application/json":  {
                                                                                                                                    "schema":  {
                                                                                                                                                   "type":  "object",
                                                                                                                                                   "properties":  {
                                                                                                                                                                      "token":  {
                                                                                                                                                                                    "type":  "string"
                                                                                                                                                                                },
                                                                                                                                                                      "user":  {
                                                                                                                                                                                   "$ref":  "#/components/schemas/User"
                                                                                                                                                                               }
                                                                                                                                                                  }
                                                                                                                                               }
                                                                                                                                }
                                                                                                       }
                                                                                       },
                                                                               "400":  {
                                                                                           "$ref":  "#/components/responses/BadRequest"
                                                                                       },
                                                                               "409":  {
                                                                                           "$ref":  "#/components/responses/Conflict"
                                                                                       }
                                                                           }
                                                         }
                                            },
                  "/api/v1/auth/login":  {
                                             "post":  {
                                                          "summary":  "Login user",
                                                          "requestBody":  {
                                                                              "required":  true,
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "type":  "object",
                                                                                                                                      "required":  [
                                                                                                                                                       "email",
                                                                                                                                                       "password"
                                                                                                                                                   ],
                                                                                                                                      "properties":  {
                                                                                                                                                         "email":  {
                                                                                                                                                                       "type":  "string",
                                                                                                                                                                       "format":  "email"
                                                                                                                                                                   },
                                                                                                                                                         "password":  {
                                                                                                                                                                          "type":  "string"
                                                                                                                                                                      }
                                                                                                                                                     }
                                                                                                                                  }
                                                                                                                   }
                                                                                          }
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "Login successful",
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "type":  "object",
                                                                                                                                                "properties":  {
                                                                                                                                                                   "token":  {
                                                                                                                                                                                 "type":  "string"
                                                                                                                                                                             },
                                                                                                                                                                   "user":  {
                                                                                                                                                                                "$ref":  "#/components/schemas/User"
                                                                                                                                                                            }
                                                                                                                                                               }
                                                                                                                                            }
                                                                                                                             }
                                                                                                    }
                                                                                    },
                                                                            "401":  {
                                                                                        "$ref":  "#/components/responses/Unauthorized"
                                                                                    }
                                                                        }
                                                      }
                                         },
                  "/api/v1/users/me":  {
                                           "get":  {
                                                       "summary":  "Get current user",
                                                       "security":  [
                                                                        {
                                                                            "bearerAuth":  [

                                                                                           ]
                                                                        }
                                                                    ],
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "User data",
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/User"
                                                                                                                                         }
                                                                                                                          }
                                                                                                 }
                                                                                 },
                                                                         "401":  {
                                                                                     "$ref":  "#/components/responses/Unauthorized"
                                                                                 }
                                                                     }
                                                   }
                                       },
...
              }
}